NET CRACKER PROJECT. Version: 0.1.111 Branch: reactev_refactor Date: 27.06.2017 Author: Gvozd
Working: project is good
Changes: sending mail from client
Next: next dao+service+controllers

Comment: MY DI*K IS LONGER THAT AN EIFEL TOWER!

+++ src/main/java/projectpackage/service/fileservice/mails/CustomMailSender.java
@@ -32,7 +32,7 @@ public class CustomMailSender implements Runnable {
             log.error(e);
             e.printStackTrace();
         } finally {
-            pdfService.deletePDF(filePath);
+            if (null!=pdfService) pdfService.deletePDF(filePath);
         }
     }
